+++ Configuration/Keybinds.cs
@@ -67,7 +67,7 @@ namespace ShipMaid
 
 		private static void OnShipMaidClosetCleanupCalled(CallbackContext context)
 		{
-			if ((Object)(object)localPlayerController == null || !localPlayerController.isPlayerControlled || localPlayerController.isTypingChat || localPlayerController.inTerminalMenu || localPlayerController.IsServer && !localPlayerController.isHostPlayerObject)
+			if ((Object)(object)localPlayerController == null || !localPlayerController.isPlayerControlled || localPlayerController.isTypingChat || localPlayerController.inTerminalMenu || localPlayerController.IsServer && !localPlayerController.isHostPlayerObject || ConfigSettings.UseOnlyTerminal.Key.Value == "Enabled")
 			{
 				return;
 			}
@@ -95,7 +95,7 @@ namespace ShipMaid
 
 		private static void OnShipMaidShipCleanupCalled(CallbackContext context)
 		{
-			if ((Object)(object)localPlayerController == null || !localPlayerController.isPlayerControlled || localPlayerController.isTypingChat || localPlayerController.inTerminalMenu || localPlayerController.IsServer && !localPlayerController.isHostPlayerObject)
+			if ((Object)(object)localPlayerController == null || !localPlayerController.isPlayerControlled || localPlayerController.isTypingChat || localPlayerController.inTerminalMenu || localPlayerController.IsServer && !localPlayerController.isHostPlayerObject || ConfigSettings.UseOnlyTerminal.Key.Value == "Enabled")
 			{
 				return;
 			}